# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Eduscope core-api — public REST surface
  version: 1.0.0
  description: |
    Contract v1 for the Eduscope UMS rewrite ("Unistream"). The Phase-2 frontend
    is built against this contract on a mock adapter before any backend exists
    (PRD §9); the Phase-4 backend implements it unchanged or bumps the version.

    **v1.0.0 (2026-08-14) — Prompt-12 drift reconciliation.** Ratification bump,
    not a breaking reshape: every ratified `x-decision` placeholder was confirmed
    as-is (PM ratification 2026-08-12, ADRs 005–012). See
    docs/design/contract-drift-report.md. Only shape change: an additive,
    optional `x-eduscope-contract` header on the quiz-sync ops (DR-10). `D-02b`
    (institute upload payload) stays deferred and keeps its tag. No consumer-breaking
    change beyond what 0.x already shipped.

    Sources of truth: docs/design/domain-model.md (entity naming — every schema
    below `$ref`s those names exactly), docs/design/state-machines.md
    (transitions R-/RA-/CH-/Q-/U-/UP-/Z-/HL-xx cited throughout), docs/PRD.md
    (screen needs LP-/AD-/QZ-/PF-xx), revamp-guide/reference/target-architecture.md §2.

    ## Conventions

    - **Commands never return final state synchronously** (single writer,
      target-architecture §3.1; rule SM-R-2). Every state-machine command
      returns `202 Accepted` + `CommandAccepted`; the resulting transition
      arrives over the WS event channel (contracts/events.md). If no event
...

...

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /recordings/{recordingId}/files/{fileId}/media:
    get:
      tags: [recordings]
      operationId: getRecordingMedia
      summary: Authenticated playback/download (LP-10) — closes B-37's open media route
      description: |
        Supports HTTP Range for playback. `?download=1` adds a
        Content-Disposition attachment header. Every request is
        authenticated + authorization-checked (INV-RC-6, playbackAuthRequired).
      parameters:
        - { $ref: '#/components/parameters/recordingId' }
        - { $ref: '#/components/parameters/fileId' }
        - name: download
          in: query
          schema: { type: boolean, default: false }
      responses:
        '200':
          description: Media bytes.
          content:
            video/mp4: { schema: { type: string, format: binary } }
            video/mp2t: { schema: { type: string, format: binary } }
        '206': { description: Partial content (Range). }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
  /recordings/{recordingId}/thumbnail.jpg:
    get:
      tags: [recordings]
      operationId: getRecordingThumbnail
      summary: Authenticated cached library thumbnail from the 10-second mark
      description: |
        Returns a compact JPEG generated from the primary derived recording at
        10 seconds. Recordings shorter than 10 seconds use their midpoint.
        Generation is serialized and cached on disk to protect the device CPU.
      parameters: [{ $ref: '#/components/parameters/recordingId' }]
      responses:
        '200':
          description: Cached recording thumbnail.
          content:
            image/jpeg: { schema: { type: string, format: binary } }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }

  # ─────────────────────────── USB export ───────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived access token (PF-17); refresh via /auth/refresh. Carries only user id + role (INV-AS-1).
    deviceAuth:
      type: http
      scheme: bearer
      description: |
        Provisioned device credential for quiz-sync paths. **v1.0.0 (DR-03):** the
        scheme is a per-device **static bearer**, minted by the deploy layer at
        provisioning (D-20) and stored hashed at rest on the quiz server;
        HMAC-signed requests are the SSO-era upgrade (D-02b / roster era), not v1.
  parameters:
...
    recordingId:
      name: recordingId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/Ulid' }
    fileId:
      name: fileId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/Ulid' }
...
  responses:
...
    Problem:
      description: Named error (never a silent no-op — R-04, INV-SB-3).
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
...
    Ulid:
      type: string
      pattern: '^[0-9A-HJKMNP-TV-Z]{26}$'
      description: Opaque ULID (INV-G-1/2, DM-13). No logic may depend on adjacency or ordering.
...
    Problem:
      type: object
      required: [status, code, title]
      properties:
        status: { type: integer }
        code:
          type: string
          description: Stable machine code — see packages/shared ErrorCode for the closed list.
          enum:
            - auth.invalid-credentials
            - auth.account-disabled
            - auth.session-revoked
            - auth.password-reset-required
            - not-authorized
            - not-found
            - validation.invalid
            - conflict
            - recorder.busy
            - storage.critical
            - provisioning.incomplete
            - volume.unavailable
            - config.invalid
            - session.not-active
            - question.immutable
            - quiz.unavailable
            - ai.unavailable
            - poweroff.refused
            - format.refused
            - export.invalid-target
            - export.insufficient-space
            - upload.not-requeueable
            - import.rejected
        title: { type: string }
        detail: { type: string }
        meta:
          type: object
          description: |
            Named-reason detail (e.g. unbound roleId; recorder owner for LP-6).
            Never secrets (INV-ST-1). Still open-ended; `reason` is the one key
            with a declared shape (v0.2, CG-11 / S01-D-5).
          properties:
            reason:
              allOf:
                - $ref: '#/components/schemas/SessionRevokedReason'
              description: |
                Set on `auth.session-revoked` only, and set on every occurrence
                of it — it is what lets a client word "your session ended"
                correctly instead of vaguely (S-01 `session expired`). `takeover`
                is R-21's `AuthSession.revokedReason`; S-06 reads the same
                vocabulary on the other side of that event.
          additionalProperties: true

    # ── enums (closed sets — adding a value is a contract bump) ──
    SessionRevokedReason:
      type: string
      enum: [expired, logout, takeover, admin]
      description: |
        Why an AuthSession stopped being valid — the value of
        `Problem.meta.reason` on `auth.session-revoked`. Mirrors
        `AuthSession.revokedReason` (domain model), so `takeover` here and
        R-21's takeover are the same fact.
...